resolve some pr comments
